Pandas 2.0 deprecation fixes and tabulate 0.8.9 update
updated the deprecated keyword level in .sum(level=1) so pandas 2.0 can work with this.
tabulate 0.8.7 needs to be updated but I haven't changed the dependencies for that because I don't know what else that will do
df.plot(sort_columns=True) is also deprecated, I've replaced it by sorting the dataframe by
df=df.sort_values(by['Nexp'])
prior to running df.plot this is bad practice and I haven't fully checked the results but hey, it works for me for now.

I'm not sure I want to be in charge of this but it's a cool little tool.