multiprocessing fixes
Print a slightly useful error message as occasionally a runsize will always produce a design will always fail, so you know what the problem is and how to fix it by setting or changing nrun_min or modifying the design file spreadsheet slightly.
added nrun_min as an option partially to address the the first problem, also can start a design with a larger runsize.
delta_nrun works now, not sure if that was in this version or the last..
cut back on screen output for the multiprocessing so its not as cluttered.