Changes to accommodate OApackage fix for array lengths >500, clean up screen output
OApackage from 2.7.7 now supports experiment runs > 500
reduce complexity of outputs to allow for evaluation of multiprocessing runs
delta_nrun is accepted as of the previous push
nrun_min also needed as an input now as low starting values <6 or 7 in tests) don't have a valid result and cause an error.
This could be fixed in code but adjusting the number of levels (increasing the smallest by 1) or just setting the starting point for nrun fixes this and addes the option for examining the window around a set of values with a bit more detail if needed.
Updated the print outputs to make it simpler to evaluate the outputs while in progress.
We now only see the runsize, the number of iterations that Doptimize has generated within the maxtime and the number of optimized arrays it's generated in that window, the orthogonality of the output and the 2 level balance.